+++ Changelog.md
@@ -2,6 +2,9 @@
 
 ### Ignition Rendering 3.X
 
+1. Skeletons for Ogre2
+    * [Pull request 224](https://bitbucket.org/ignitionrobotics/ign-rendering/pull-requests/224)
+
 ### Ignition Rendering 3.0.0 (2019-12-10)
 
 1. Support setting cast shadows